feat: id에 해당하는 레코드가 존재하는지 판별하는 기능 추가
- delete시 이를 통해 검증하도록 수정

+++ src/main/java/roomescape/RoomRepository.java
@@ -65,6 +65,10 @@ public class RoomRepository {
 		}
 	}
 
+	public boolean existsById(Long id) {
+		return jdbcTemplate.queryForObject("select exists(select * from reservation where id = ? limit 1)", Boolean.class, id);
+	}
+
 	public List<Room> findAll() {
 		return jdbcTemplate.query(
 				"select id, name, date, time from reservation",
@@ -80,7 +84,9 @@ public class RoomRepository {
 	}
 
 	public void deleteById(Long id) {
-		findById(id);
+		if(!existsById(id)) {
+			throw new DomainEmptyFieldException("해당하는 예약이 없습니다.");
+		}
 
 		jdbcTemplate.update("delete from reservation where id = ?", id);
 	}
